Changing from custom Dataset object to new object (`hypers.hparray`)
Subclasses numpy.ndarray and extends functionality for
hyperspectral data.
