+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Version 0.0.12 (pre-release)
 #### Features
-+ Added figsize argument to relevant methods in Dataset
++ `Dataset` object has been removed in favour of subclassing the numpy array object `numpy.ndarray` 
+and extending its functionality for hyperspectral data. This forms the main data type for hypers and 
+is `hypers.hparray`.
 
 ## Version 0.0.11 (pre-release)
 #### Features
