+++ hypers/core/__init__.py
@@ -1,0 +1,1 @@
+from hypers.core.hparray import hparray
