+++ hypers/exceptions/__init__.py
@@ -1,0 +1,11 @@
+"""
+Exposes public exceptions and warinings
+"""
+
+
+class DimensionError(ValueError):
+    """
+    Incorrect number of dimensions. Should be raised when an
+    operation requires a certain number of dimensions and this
+    requirement is not met.
+    """
