+++ hypers/learning/__init__.py
@@ -1,6 +1,4 @@
-from hypers.learning._cluster import _data_cluster
-from hypers.learning._decompose import _decompose, _pca, _vca, _ica, _nmf
-from hypers.learning._abundance import _ucls
-from hypers.learning._mixture import _data_mixture
+# from hypers.learning._abundance import _ucls
+# from hypers.learning._mixture import _data_mixture
 
-__all__ = ['_data_cluster', '_vca', '_ucls', '_data_mixture', '_pca', '_ica', '_nmf']
+# __all__ = ['_kmeans', '_vca', '_ucls', '_data_mixture', '_pca', '_ica', '_nmf']
