+++ hypers/plotting/view/__init__.py
@@ -1,0 +1,3 @@
+from hypers.plotting.view._hsiDialog import hsiPlot
+
+__all__ = ['hsiPlot']
