--- hypers/tools/_empty.py
@@ -1,5 +1,0 @@
-# Provides an empty class
-
-
-class null:
-    pass
