--- hypers/view/__init__.py
@@ -1,3 +1,0 @@
-from hypers.view._hsiDialog import hsiPlot
-
-__all__ = ['hsiPlot']
