--- hypers/view/_form/mainwindow.ui
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MainWindow</class>
- <widget class="QMainWindow" name="MainWindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>1243</width>
-    <height>458</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>View Hyperspectral Image</string>
-  </property>
-  <property name="styleSheet">
-   <string notr="true">%background: #161616</string>
-  </property>
-  <widget class="QWidget" name="centralwidget">
-   <layout class="QGridLayout" name="gridLayout" rowstretch="0,95,5">
-    <item row="1" column="0">
-     <layout class="QHBoxLayout" name="horizontalLayout_2" stretch="40,60">
-      <item>
-       <widget class="ImageView" name="imagewin"/>
-      </item>
-      <item>
-       <widget class="PlotWidget" name="specwin"/>
-      </item>
-     </layout>
-    </item>
-    <item row="2" column="0">
-     <layout class="QHBoxLayout" name="horizontalLayout_3">
-      <item>
-       <widget class="QSlider" name="slider">
-        <property name="maximum">
-         <number>10</number>
-        </property>
-        <property name="pageStep">
-         <number>1</number>
-        </property>
-        <property name="orientation">
-         <enum>Qt::Horizontal</enum>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QPushButton" name="updateSpectrum">
-        <property name="text">
-         <string> Update Spectrum </string>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QPushButton" name="updateImage">
-        <property name="text">
-         <string> Update Image </string>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QPushButton" name="Reset">
-        <property name="text">
-         <string>RESET</string>
-        </property>
-       </widget>
-      </item>
-     </layout>
-    </item>
-   </layout>
-  </widget>
- </widget>
- <customwidgets>
-  <customwidget>
-   <class>ImageView</class>
-   <extends>QGraphicsView</extends>
-   <header>pyqtgraph</header>
-  </customwidget>
-  <customwidget>
-   <class>PlotWidget</class>
-   <extends>QGraphicsView</extends>
-   <header>pyqtgraph</header>
-  </customwidget>
- </customwidgets>
- <resources/>
- <connections/>
-</ui>
